requirements.txt
Flask
firebase-admin
python-dotenv
google-auth
google_auth_oauthlib